new-modules compile with ifort in moderate times
Some modules (symmetry, ff_build, internal, zmat_manage)
took veeery long to be compiled with ifort. The issue
seems to be related to the use of str_resmol within the
modules as auxiliary variables (maybe due to the huge size)
So, the modules were changed replacing the use of the
auxiliar molecules to scalars or arrays of real(8) that
play the same role.

Relevant changes:
* symmetry: the geometry is changed place the molec into
  std orientation. Now, the coodinates are stored first in
  arrays(X0(:), Y0(:), Z0(:)) dinamically allocated, and
  restored to molec before leaving the subroutines
  The _old subroutine is commeted out
* zmat_manage: str_resmol was declared in rmsd_fit routines
  but not used (so it is not deffined now)
* internal: str_resmol was used in NumBder, as an auxiliar
  to perform the numerical diplacements. Now, the displacement
  are applied into the original molec, but the input geometry
  is first stored into X0, Y0 and Z0 scalars and restored
  each step
* ff_build: in this case, there does not seem to be any call
  to auxiliar str_resmol. The part of the module connected to
  build_top things is commented out, leaving only gen_bonded()
  This routine should be accomodated somewhere else