Improve error handling in readers
When the eror_lable is present the error is bypassed
with the error info included in the lable. The main
program should manage the error. This is useful for
trajectory reader till the end of the file (rises error)
and would be used with fchk2gro_trj.f90

IT IS STILL UNDER DEVELOPMENT. THIS COMMIT IS NOT PRODUCTION READY