Fix mess with projections in nm_internal
In commit c55d599 the method to manage "-intmode sel -intfile FILE"
was changed to test the projections. The modifications were left
but not documented. This broke the consistency between the different
tools to manage the internal sets. Since the projection is equivalent
to the method used uo to now (use reduced B to generate Hint), that
standard method is left. The projection is left with -prj-trS is
activated. Concretely, if so, the input internals are used to generate
the projection, the all calculations are done with the whole space Ball
but Hx is projected with P=B^+B (B defined in the input set)