Implemented changes suggested by lf95 compiler
After a test compilation with lf95, some conflicts with the
fortran standard are detected. We here solve just some of them

One of the changes is reducing the sizes of the atomic arrays
in structure_types. Note that the MAX_ATM_RES need to be at
least 600 to comply with the NDIM used in the code.
Dynamic memory allocation is needed

Support for the lf95 compiler is also added. The check for the
-cpp flags is disabled (it does not work properly)