Add -prjS-c options to internal_duschinsky
Follwing the simplified approach introduced in nm_internal.
Since we have observed the equivalence between the different
approaches, some options are no longer supported (were useful
for testing, but not for "final user")

In this version, the option -prjS-c is working properly within
internal_duschinsky

The source is in *_prjCrt.f90, while keeping the legacy code
in the original file for future reference.