Version of nm_internal that mix projecition&red space
Apply the projection to Hx with one set and then perform
the vibrational analysis with that Hx and using another
set to construct the reduced space. For instance, we can
project out a torsion, and then work in a space that contains
all but torsion.

This version allows combinations both in the projection (as usual)
and in the construction of the internal working space (this is new)