Bugfix in read_top (affects gen_oniom)
The read_top subroutine now reads the residues as they are found
in the top. This is bad, becouse some moltypes defined in the top
are not in the molecules section. A smarter way to read the top
is needed. In the meantime, support is provided to skip reading
moltypes that are not in the system (but defined in the top).
This solves an issue, as the read_top routine was trying to read
molecules that were not present (e.g. ions), which led to a segfault

The present bugfix solves the problem. But still the top reader should
be improved